+++ routes.go
@@ -22,6 +22,8 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/justinas/alice"
+	"github.com/xmidt-org/arrange"
+	"github.com/xmidt-org/arrange/arrangehttp"
 	"github.com/xmidt-org/httpaux"
 	"github.com/xmidt-org/touchstone/touchhttp"
 	"go.uber.org/fx"
@@ -30,7 +32,7 @@ import (
 type PrimaryRouterIn struct {
 	fx.In
 	Router    *mux.Router `name:"server_primary"`
-	AuthChain alice.Chain `name:"primary_auth_chain"`
+	AuthChain alice.Chain `name:"auth_chain"`
 }
 
 type MetricsRoutesIn struct {
@@ -55,19 +57,57 @@ type MetricMiddlewareOut struct {
 	Health  alice.Chain `name:"middleware_health_metrics"`
 }
 
-func buildPrimaryRoutes(in PrimaryRouterIn) {
+func handlePrimaryEndpoint(in PrimaryRouterIn) {
 	in.Router.Use(in.AuthChain.Then)
 	in.Router.PathPrefix("/").Handler(httpaux.ConstantHandler{StatusCode: http.StatusOK})
 }
 
-func buildMetricsRoutes(in MetricsRoutesIn) {
-	if in.Handler != nil {
-		in.Router.Handle("/metrics", in.Handler).Methods("GET")
-	}
+func handledMetricEndpoint(in MetricsRoutesIn) {
+	in.Router.Handle("/metrics", in.Handler).Methods("GET")
 }
 
 func metricMiddleware(bundle touchhttp.ServerBundle) (out MetricMiddlewareOut) {
 	out.Primary = alice.New(bundle.ForServer("server_primary").Then)
 	out.Health = alice.New(bundle.ForServer("server_health").Then)
 	return
+}
+
+func provideServers() fx.Option {
+	return fx.Options(
+		fx.Provide(
+			metricMiddleware,
+		),
+		arrangehttp.Server{
+			Name: "server_primary",
+			Key:  "servers.primary",
+			Inject: arrange.Inject{
+				PrimaryMMIn{},
+			},
+		}.Provide(),
+		arrangehttp.Server{
+			Name: "server_health",
+			Key:  "servers.health",
+			Inject: arrange.Inject{
+				HealthMMIn{},
+			},
+			Invoke: arrange.Invoke{
+				func(r *mux.Router) {
+					r.Handle("/health", httpaux.ConstantHandler{
+						StatusCode: http.StatusOK,
+					}).Methods("GET")
+				},
+			},
+		}.Provide(),
+		arrangehttp.Server{
+			Name: "server_metrics",
+			Key:  "servers.metrics",
+		}.Provide(),
+		fx.Invoke(
+			serverValidator{Key: "servers.primary"}.Validate,
+			serverValidator{Key: "servers.metrics"}.Validate,
+			serverValidator{Key: "servers.health"}.Validate,
+			handlePrimaryEndpoint,
+			handledMetricEndpoint,
+		),
+	)
 }