+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- Upgrade bascule and rm undesired deps like argus, webpa-common and themis. [#13](https://github.com/xmidt-org/authbaton/pull/13)
 
 ## [v0.1.0]
 - Start migrating away from Themis sub-packages. [#6](https://github.com/xmidt-org/authbaton/pull/6)